refactor: 标准插件代码目录规范化
