+++ pipeline_plugins/components/collections/sites/open/nodeman/__init__.py
@@ -11,5 +11,4 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express o
 specific language governing permissions and limitations under the License.
 """
 
-from .create_task.nodeman import NodemanCreateTaskService  # noqa
-from .create_task.nodeman import NodemanCreateTaskComponent  # noqa
+from .create_task.legacy import NodemanCreateTaskService, NodemanCreateTaskComponent  # noqa
