+++ pipeline_plugins/components/collections/sites/open/nodeman/create_task/legacy.py
@@ -306,7 +306,4 @@ class NodemanCreateTaskComponent(Component):
     name = _("新建任务")
     code = "nodeman_create_task"
     bound_service = NodemanCreateTaskService
-    form = "%scomponents/atoms/sites/%s/nodeman/create_task/nodeman_create_task.js" % (
-        settings.STATIC_URL,
-        settings.RUN_VER,
-    )
+    form = "%scomponents/atoms/nodeman/create_task/legacy.js" % settings.RUN_VER
