+++ pipeline_plugins/components/collections/sites/open/nodeman/create_task/v2_0.py
@@ -374,8 +374,5 @@ class NodemanCreateTaskComponent(Component):
     name = _("新建任务")
     code = "nodeman_create_task"
     bound_service = NodemanCreateTaskService
-    form = "%scomponents/atoms/sites/%s/nodeman/create_task/v2_0.js" % (
-        settings.STATIC_URL,
-        settings.RUN_VER,
-    )
+    form = "%scomponents/atoms/nodeman/create_task/v2_0.js" % settings.STATIC_URL
     version = VERSION
