+++ pipeline_plugins/components/static/components/atoms/nodeman/create_task/legacy.js
