+++ pipeline_plugins/components/static/components/atoms/nodeman/create_task/v2_0.js
