--- static/components/atoms/sites/open/job/job_fast_push_file.js
@@ -1,144 +1,0 @@
-/**
- * Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community
- * Edition) available.
- * Copyright (C) 2017-2020 THL A29 Limited, a Tencent company. All rights reserved.
- * Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://opensource.org/licenses/MIT
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-(function () {
-    $.atoms.job_fast_push_file = [
-        {
-            tag_code: "biz_cc_id",
-            type: "select",
-            attrs: {
-                name: gettext("业务"),
-                hookable: true,
-                remote: true,
-                remote_url: $.context.get('site_url') + 'pipeline/cc_get_business_list/',
-                remote_data_init: function (resp) {
-                    return resp.data;
-                },
-                disabled: !$.context.canSelectBiz(),
-                validation: [
-                    {
-                        type: "required"
-                    }
-                ]
-            },
-            methods: {
-                _tag_init: function () {
-                    if (this.value) {
-                        return
-                    }
-                    this._set_value($.context.getBkBizId())
-                }
-            }
-        },
-        {
-            tag_code: "job_source_files",
-            type: "datatable",
-            attrs: {
-                name: gettext("源文件"),
-                editable: true,
-                add_btn: true,
-                columns: [
-                    {
-                        tag_code: "ip",
-                        type: "input",
-                        attrs: {
-                            name: gettext("IP"),
-                            width: '100px',
-                            editable: true,
-                            validation: [
-                                {
-                                    type: "required"
-                                }
-                            ]
-                        }
-                    },
-                    {
-                        tag_code: "files",
-                        type: "textarea",
-                        attrs: {
-                            name: gettext("文件路径"),
-                            placeholder: gettext("多个用换行分隔"),
-                            width: '200px',
-                            editable: true,
-                            validation: [
-                                {
-                                    type: "required"
-                                }
-                            ]
-                        }
-                    },
-                    {
-                        tag_code: "account",
-                        type: "input",
-                        attrs: {
-                            name: gettext("执行账户"),
-                            width: '100px',
-                            editable: true,
-                            validation: [
-                                {
-                                    type: "required"
-                                }
-                            ]
-                        }
-                    }
-                ],
-                hookable: true,
-                validation: [
-                    {
-                        type: "required"
-                    }
-                ]
-            }
-        },
-        {
-            tag_code: "job_ip_list",
-            type: "textarea",
-            attrs: {
-                name: gettext("目标IP"),
-                placeholder: gettext("IP必须填写【云区域ID:IP】或者【IP】格式之一，多个用换行分隔；【IP】格式需要保证所填写的内网IP在配置平台(CMDB)的该业务中是唯一的"),
-                hookable: true,
-                validation: [
-                    {
-                        type: "required"
-                    }
-                ]
-            }
-        },
-        {
-            tag_code: "job_account",
-            type: "input",
-            attrs: {
-                name: gettext("目标账户"),
-                placeholder: gettext("请输入在蓝鲸作业平台上注册的账户名"),
-                hookable: true,
-                validation: [
-                    {
-                        type: "required"
-                    }
-                ]
-            }
-        },
-        {
-            tag_code: "job_target_path",
-            type: "input",
-            attrs: {
-                name: gettext("目标路径"),
-                placeholder: gettext("请输入绝对路径"),
-                hookable: true,
-                validation: [
-                    {
-                        type: "required"
-                    }
-                ]
-            }
-        }
-    ]
-})();
